feat: collision handlers with pymunk and projectile collision
may contain bugs
